I moved the props check to error file of SelectClass
